projects/capstone/starter/requirements.txt
alembic==1.4.2
astroid==2.2.5
attrs==19.3.0
awscli==1.18.31
beautifulsoup4==4.8.2
botocore==1.15.31
cachetools==4.0.0
certifi==2019.11.28
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
docutils==0.15.2
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
future==0.17.1
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.5
lazy-object-proxy==1.4.0
libusb==1.0.22b2
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.3
Pillow==7.0.0
pluggy==0.13.1
protobuf==3.11.3
psycopg2-binary==2.8.5
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.3.1
PyJWT==1.7.1
pylint==2.3.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
PyQt5==5.14.1
PyQt5-sip==12.7.1
pytest==5.4.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose-cryptodome==1.3.2
pytz==2019.3
pyusb==1.0.2
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.3.3
six==1.12.0
soupsieve==2.0
SQLAlchemy==1.3.3
tqdm==4.43.0
typed-ast==1.3.5
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==0.15.2
wrapt==1.11.1
zipp==3.1.0
